[FEATURE] Url rewriting of event to insert event title
