+++ collectives/helpers.py
@@ -1,8 +1,6 @@
 """Functions to help on standard taks
 """
 from datetime import datetime
-import re
-import unicodedata
 
 from dateutil import tz
 
@@ -18,20 +16,3 @@ def current_time():
     # For Python to allow comparisons we need to strip the
     # tz information from our local time
     return now.replace(tzinfo=None)
-
-
-_slugify_strip_re = re.compile(r"[^\w\s-]")
-_slugify_hyphenate_re = re.compile(r"[-\s]+")
-
-
-def slugify(value):
-    """String normalisation.
-
-    Normalizes string, converts to lowercase, removes non-alpha characters,
-    and converts spaces to hyphens.
-
-    From Django's "django/template/defaultfilters.py".
-    """
-    value = unicodedata.normalize("NFKD", value).encode("ascii", "ignore")
-    value = _slugify_strip_re.sub("", value.decode("ascii")).strip().lower()
-    return _slugify_hyphenate_re.sub("-", value)
